Fixes regex to recognise JDKs without .0 version
This script assumes that all JDK's have a `major.0.patch` version number. This is true for many of them, and probably most of the ones that are used the most. But there are also JDK's like `15.ea.36` (in Java.net) and `20.2.0.r11` (in GraalVM). This fix makes sure that `j` can switch to those, too.